Add files via upload
Break a cryptographic system composed of the two Python scripts
EncryptForFun.py and DecryptForFun.py described in Section 2.11 in Lecture 2. As you will recall, the script EncryptForFun.py can be used for
encrypting a message file while the script DecryptForFun.py recovers that
message from the ciphertext produced with the previous script.